[bugfix]modal对style配置支持的bug(可以复写min-height等)
